Add new function xfce_spawn_on_screen_with_child_watch().
New function to spawn a command with child watch. For the callback
a closure is used (void__int). Although the PID is not returned,
this is good enough for most implementations.

Internal change to use environ again instead of the glib code,
which only resulted in a lot of string copying.

Use seconds timeout to the startup expire time.