Restore vendor information handling.
Restore the configure flags and simplify the vendor file
handing. The file is not name specific anymore, but just
stored in $prefix/share/xfce4/vendorinfo.
To hide the message the file is not found, packages
can 'touch' the file, so it has a zero length.

Note that --with-vendor-info "Your Name" is also supported.