Fix alt + print being detected as SysReq (bug #7897).
If we were able to grab it, it means that this combination is not used
as SysReq, users should then be able to use that.

Note that this won't work by default on most distributions because the
SysReq key is enabled and X won't pass the event. You need to disable the
SysReq key, for example by setting "kernel.sysrq = 0" in /etc/sysctl.conf.