Add support for GOBJECT-INTROSPECTION (Bug #13461)
GOjbect Introspection allows various other languages to automatically
generate bindings they can use. For details see:
https://wiki.gnome.org/action/show/Projects/GObjectIntrospection

To support this, we need to add a gobject and introspection
dependency. The introspection one is optional and it is what
determines if the libxfce4ui-2 gir is created.
Only libxfce4ui-2 is created (the Gtk3 version) as it makes little
sense to support new applications in Gtk2.