Add a dialog to confirm closure of multiple tabs (bug #15873)
Adds a new function xfce_dialog_confirm_close_tabs, which runs a dialog
box to
ask the user if they wish to close the current tab, close the whole
window, or
cancel. The function has the option of adding a check box to the dialog
to let
the user choose whether they wish to see this dialog in future.