XfceTitledDialog: Move buttons from headerbar to actionarea
Instead of letting all buttons automatically end up in the dialog's
headerbar we want to preserve the previous behavior of
xfce_titled_dialog_new_with_mixed_buttons and keep buttons in the action
area.

This change breaks the functionality of gtk_dialog_set_default_response
for which a replacement will have to be introduced.