xfce-dialogs: Drop default window titles
With client side decorations this looks weird anyway plus the titles are
not very telling. If the creator of a dialog wants to set a meaningful
title, it will still be displayed.
Another use-case for window titles would be the alt-tab dialog or any
other type of window list, but GtkMessageDialogs don't show up there
anyway.