XfceTitledDialog: Keep buttons in the action area
Previously XfceTitledDialog would end up having its buttons moved to the
headerbar by the constructed function of GtkDialog. By blocking this
function we ensure that we have
1) a headerbar and
2) an action area.

This is only desirable for XfceTitledDialog, which is a special kind of
dialog (settings dialog).

This commit fixes the behavior above when using the class directly.