Add a widget for filename input (Bug #16542)
Adds a new widget, XfceFilenameInput, which essentially provides a smart
version of the GtkEntry widget especially for entering filenames for
creating
or renaming files. In such situations it is necessary to check the
filename
the user has entered for validity (for example, it cannot be too long or
contain directory separator characters). The new widget provides this,
with
as-you-type error messages to tell the user if the filename they have
entered
is not valid.

Co-authored-by: Alexander Schwinn <alexxcons@xfce.org>